requirements.txt
asgiref==3.7.2
chardet==5.2.0
Django==3.2
django-excel==0.0.10
django-excel-response==2.0.5
django-filter==23.5
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
jdcal==1.4.1
lml==0.1.0
numpy==1.26.4
openpyxl==2.4.11
pandas==2.2.0
pillow==10.2.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
texttable==1.7.0
typing_extensions==4.9.0
tzdata==2024.1
XlsxWriter==3.2.0
xlwt==1.3.0
